fix(接口测试):  csv文件类型限制&场景步骤调整
