+++ frontend/src/views/api-test/scenario/components/common/customApiDrawer.vue
@@ -30,6 +30,7 @@
               {{ title }}
             </div>
             <MsIcon
+              v-if="!props.step || !props.step.isQuoteScenarioStep"
               type="icon-icon_edit_outlined"
               class="cursor-pointer hover:text-[rgb(var(--primary-5))]"
               @click="isShowEditStepNameInput = true"
